refactor: use global_asm! instead of standalone asm file
Move asm.S to src/asm.S, and use global_asm! and include_str!
to compile it.

Remove unnecessary pre-compiled static library files to simplify
the commit process.

Remove unnecessary build dependency riscv-target.

Add cache control, currenrtly support enable and disable I-Cache
and D-Cache.
