+++ .gitignore
@@ -3,3 +3,4 @@ target/
 bin/*.after
 bin/*.before
 bin/*.o
+.DS_Store
