+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "hpm-rt"
 version = "0.1.0"
 rust-version = "1.59"
-repository = "https://github.com/tfx2001/hpm-rt"
+repository = "https://github.com/hpm-rs/hpm-rt"
 authors = ["tfx2001 <tfx2001@outlook.com>"]
 categories = ["embedded", "no-std"]
 description = "Minimal runtime / startup for HPMicro MCUs"
@@ -21,6 +21,3 @@ riscv-rt-macros = { path = "macros", version = "0.2.0" }
 
 [dev-dependencies]
 panic-halt = "0.2.0"
-
-[build-dependencies]
-riscv-target = "0.1.2"
