+++ build.rs
@@ -1,26 +1,9 @@
-// NOTE: Adapted from cortex-m/build.rs
-extern crate riscv_target;
-
-use riscv_target::Target;
 use std::env;
 use std::fs;
 use std::path::PathBuf;
 
 fn main() {
-    let target = env::var("TARGET").unwrap();
     let out_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
-    let name = env::var("CARGO_PKG_NAME").unwrap();
-
-    if target.starts_with("riscv") {
-        let mut target = Target::from_target_str(&target);
-        target.retain_extensions("imfdc");
-        let archive: String;
-        archive = format!("bin/{}.a", target.to_string());
-
-        fs::copy(&archive, out_dir.join(format!("lib{}.a", name))).unwrap();
-        println!("cargo:rerun-if-changed={}", archive);
-        println!("cargo:rustc-link-lib=static={}", name);
-    }
 
     // Put the linker script somewhere the linker can find it
     if cfg!(feature = "flash-xip") {
