+++ src/lib.rs
@@ -322,276 +322,13 @@
 //! ```
 //!
 //! Default implementation of this function stucks in a busy-loop.
-//!
-//! # Features
-//!
-//! ## `s-mode`
-//!
-//! The supervisor mode feature (`s-mode`) can be activated via [Cargo features](https://doc.rust-lang.org/cargo/reference/features.html).
-//!
-//! For example:
-//! ``` text
-//! [dependencies]
-//! riscv-rt = {features=["s-mode"]}
-//! ```
-//! Internally, riscv-rt uses different versions of precompiled static libraries
-//! for (i) machine mode and (ii) supervisor mode. If the `s-mode` feature was activated,
-//! the build script selects the s-mode library. While most registers/instructions have variants for
-//! both `mcause` and `scause`, the `mhartid` hardware thread register is not available in supervisor
-//! mode. Instead, the hartid is passed as parameter by a bootstrapping firmware (i.e., SBI).
-//!
-//! Use case: QEMU supports [OpenSBI](https://github.com/riscv-software-src/opensbi) as default firmware.
-//! Using the SBI requires riscv-rt to be run in supervisor mode instead of machine mode.
-//! ``` text
-//! APP_BINARY=$(find target -name app)
-//! sudo qemu-system-riscv64 -m 2G -nographic -machine virt -kernel $APP_BINARY
-//! ```
-//! It requires the memory layout to be non-overlapping, like
-//! ``` text
-//! MEMORY
-//! {
-//!   RAM : ORIGIN = 0x80200000, LENGTH = 0x8000000
-//!   FLASH : ORIGIN = 0x20000000, LENGTH = 16M
-//! }
-//! ```
-
-// NOTE: Adapted from cortex-m/src/lib.rs
 #![no_std]
+// NOTE: Adapted from cortex-m/src/lib.rs
 #![deny(missing_docs)]
 
-#[cfg(feature = "s-mode")]
-use riscv::register::{scause as xcause, stvec as xtvec, stvec::TrapMode as xTrapMode};
+/// L1-Cache control
+pub mod cache;
 
-#[cfg(not(feature = "s-mode"))]
-use riscv::register::{mcause as xcause, mhartid, mtvec as xtvec, mtvec::TrapMode as xTrapMode};
+mod startup;
 
-pub use riscv_rt_macros::{entry, pre_init};
-
-#[export_name = "error: riscv-rt appears more than once in the dependency graph"]
-#[doc(hidden)]
-pub static __ONCE__: () = ();
-
-extern "C" {
-    // Boundaries of the .bss section
-    static mut _ebss: u32;
-    static mut _sbss: u32;
-
-    // Boundaries of the .data section
-    static mut _edata: u32;
-    static mut _sdata: u32;
-
-    // Initial values of the .data section (stored in Flash)
-    static _sidata: u32;
-}
-
-/// Rust entry point (_start_rust)
-///
-/// Zeros bss section, initializes data section and calls main. This function
-/// never returns.
-#[link_section = ".init.rust"]
-#[export_name = "_start_rust"]
-pub unsafe extern "C" fn start_rust(a0: usize, a1: usize, a2: usize) -> ! {
-    #[rustfmt::skip]
-    extern "Rust" {
-        // This symbol will be provided by the user via `#[entry]`
-        fn main(a0: usize, a1: usize, a2: usize) -> !;
-
-        // This symbol will be provided by the user via `#[pre_init]`
-        fn __pre_init();
-
-        fn _setup_interrupts();
-
-        fn _mp_hook(hartid: usize) -> bool;
-    }
-
-    // sbi passes hartid as first parameter (a0)
-    #[cfg(feature = "s-mode")]
-    let hartid = a0;
-    #[cfg(not(feature = "s-mode"))]
-    let hartid = mhartid::read();
-
-    if _mp_hook(hartid) {
-        __pre_init();
-
-        r0::zero_bss(&mut _sbss, &mut _ebss);
-        r0::init_data(&mut _sdata, &mut _edata, &_sidata);
-    }
-
-    // TODO: Enable FPU when available
-
-    _setup_interrupts();
-
-    main(a0, a1, a2);
-}
-
-/// Registers saved in trap handler
-#[allow(missing_docs)]
-#[repr(C)]
-#[derive(Debug)]
-pub struct TrapFrame {
-    pub ra: usize,
-    pub t0: usize,
-    pub t1: usize,
-    pub t2: usize,
-    pub t3: usize,
-    pub t4: usize,
-    pub t5: usize,
-    pub t6: usize,
-    pub a0: usize,
-    pub a1: usize,
-    pub a2: usize,
-    pub a3: usize,
-    pub a4: usize,
-    pub a5: usize,
-    pub a6: usize,
-    pub a7: usize,
-}
-
-/// Trap entry point rust (_start_trap_rust)
-///
-/// `scause`/`mcause` is read to determine the cause of the trap. XLEN-1 bit indicates
-/// if it's an interrupt or an exception. The result is examined and ExceptionHandler
-/// or one of the core interrupt handlers is called.
-#[link_section = ".trap.rust"]
-#[export_name = "_start_trap_rust"]
-pub extern "C" fn start_trap_rust(trap_frame: *const TrapFrame) {
-    extern "C" {
-        fn ExceptionHandler(trap_frame: &TrapFrame);
-        fn DefaultHandler();
-    }
-
-    unsafe {
-        let cause = xcause::read();
-
-        if cause.is_exception() {
-            ExceptionHandler(&*trap_frame)
-        } else {
-            if cause.code() < __INTERRUPTS.len() {
-                let h = &__INTERRUPTS[cause.code()];
-                if h.reserved == 0 {
-                    DefaultHandler();
-                } else {
-                    (h.handler)();
-                }
-            } else {
-                DefaultHandler();
-            }
-        }
-    }
-}
-
-#[doc(hidden)]
-#[no_mangle]
-#[allow(unused_variables, non_snake_case)]
-pub fn DefaultExceptionHandler(trap_frame: &TrapFrame) -> ! {
-    loop {
-        // Prevent this from turning into a UDF instruction
-        // see rust-lang/rust#28728 for details
-        continue;
-    }
-}
-
-#[doc(hidden)]
-#[no_mangle]
-#[allow(unused_variables, non_snake_case)]
-pub fn DefaultInterruptHandler() {
-    loop {
-        // Prevent this from turning into a UDF instruction
-        // see rust-lang/rust#28728 for details
-        continue;
-    }
-}
-
-/* Interrupts */
-#[doc(hidden)]
-pub enum Interrupt {
-    UserSoft,
-    SupervisorSoft,
-    MachineSoft,
-    UserTimer,
-    SupervisorTimer,
-    MachineTimer,
-    UserExternal,
-    SupervisorExternal,
-    MachineExternal,
-}
-
-pub use self::Interrupt as interrupt;
-
-extern "C" {
-    fn UserSoft();
-    fn SupervisorSoft();
-    fn MachineSoft();
-    fn UserTimer();
-    fn SupervisorTimer();
-    fn MachineTimer();
-    fn UserExternal();
-    fn SupervisorExternal();
-    fn MachineExternal();
-}
-
-#[doc(hidden)]
-pub union Vector {
-    pub handler: unsafe extern "C" fn(),
-    pub reserved: usize,
-}
-
-#[doc(hidden)]
-#[no_mangle]
-pub static __INTERRUPTS: [Vector; 12] = [
-    Vector { handler: UserSoft },
-    Vector {
-        handler: SupervisorSoft,
-    },
-    Vector { reserved: 0 },
-    Vector {
-        handler: MachineSoft,
-    },
-    Vector { handler: UserTimer },
-    Vector {
-        handler: SupervisorTimer,
-    },
-    Vector { reserved: 0 },
-    Vector {
-        handler: MachineTimer,
-    },
-    Vector {
-        handler: UserExternal,
-    },
-    Vector {
-        handler: SupervisorExternal,
-    },
-    Vector { reserved: 0 },
-    Vector {
-        handler: MachineExternal,
-    },
-];
-
-#[doc(hidden)]
-#[no_mangle]
-#[rustfmt::skip]
-pub unsafe extern "Rust" fn default_pre_init() {}
-
-#[doc(hidden)]
-#[no_mangle]
-#[rustfmt::skip]
-pub extern "Rust" fn default_mp_hook(hartid: usize) -> bool {
-    match hartid {
-        0 => true,
-        _ => loop {
-            unsafe { riscv::asm::wfi() }
-        },
-    }
-}
-
-/// Default implementation of `_setup_interrupts` that sets `mtvec`/`stvec` to a trap handler address.
-#[doc(hidden)]
-#[no_mangle]
-#[rustfmt::skip]
-pub unsafe extern "Rust" fn default_setup_interrupts() {
-    extern "C" {
-        fn _start_trap();
-    }
-
-    xtvec::write(_start_trap as usize, xTrapMode::Direct);
-}
+pub use startup::{entry, Interrupt};
